refactored efilingdiligenclient to aidiligenclient
