+++ src/backend/ai-reviewer-api/src/test/java/ca/bc/gov/open/jag/efilingreviewerapi/document/documentsApiDelegateImpl/DocumentEventTest.java
@@ -3,9 +3,9 @@ package ca.bc.gov.open.jag.efilingreviewerapi.document.documentsApiDelegateImpl;
 import ca.bc.gov.open.jag.aidiligenclient.diligen.DiligenService;
 import ca.bc.gov.open.jag.aidiligenclient.diligen.model.DiligenDocumentDetails;
 import ca.bc.gov.open.jag.aidiligenclient.diligen.processor.FieldProcessor;
-import ca.bc.gov.open.jag.efilingdiligenclient.api.model.Field;
-import ca.bc.gov.open.jag.efilingdiligenclient.api.model.ProjectFieldsResponse;
-import ca.bc.gov.open.jag.efilingdiligenclient.api.model.ProjectFieldsResponseData;
+import ca.bc.gov.open.jag.aidiligenclient.api.model.Field;
+import ca.bc.gov.open.jag.aidiligenclient.api.model.ProjectFieldsResponse;
+import ca.bc.gov.open.jag.aidiligenclient.api.model.ProjectFieldsResponseData;
 import ca.bc.gov.open.jag.efilingreviewerapi.api.model.DocumentEvent;
 import ca.bc.gov.open.jag.efilingreviewerapi.document.DocumentsApiDelegateImpl;
 import ca.bc.gov.open.jag.efilingreviewerapi.document.models.Document;
