+++ src/backend/libs/ai-diligen-client-starter/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-    ca.bc.gov.open.jag.efilingdiligenclientstarter.AutoConfiguration
+    ca.bc.gov.open.jag.aidiligenclientstarter.AutoConfiguration
