+++ src/backend/libs/ai-diligen-client/src/main/java/ca/bc/gov/open/jag/aidiligenclient/diligen/mapper/DiligenDocumentDetailsMapper.java
@@ -5,9 +5,9 @@ import java.util.List;
 import org.mapstruct.Mapper;
 
 import ca.bc.gov.open.jag.aidiligenclient.diligen.model.DiligenDocumentDetails;
-import ca.bc.gov.open.jag.efilingdiligenclient.api.model.Field;
-import ca.bc.gov.open.jag.efilingdiligenclient.api.model.InlineResponse2003DataFileDetails;
-import ca.bc.gov.open.jag.efilingdiligenclient.api.model.ProjectFieldsResponse;
+import ca.bc.gov.open.jag.aidiligenclient.api.model.Field;
+import ca.bc.gov.open.jag.aidiligenclient.api.model.InlineResponse2003DataFileDetails;
+import ca.bc.gov.open.jag.aidiligenclient.api.model.ProjectFieldsResponse;
 
 
 @Mapper
