+++ src/backend/libs/ai-diligen-client/src/main/java/ca/bc/gov/open/jag/aidiligenclient/diligen/processor/FieldProcessorImpl.java
@@ -2,7 +2,7 @@ package ca.bc.gov.open.jag.aidiligenclient.diligen.processor;
 
 import ca.bc.gov.open.jag.aidiligenclient.diligen.model.DocumentConfig;
 import ca.bc.gov.open.jag.aidiligenclient.diligen.model.PropertyConfig;
-import ca.bc.gov.open.jag.efilingdiligenclient.api.model.Field;
+import ca.bc.gov.open.jag.aidiligenclient.api.model.Field;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import org.apache.commons.lang3.StringUtils;
