+++ src/backend/libs/ai-diligen-client/src/test/java/ca/bc/gov/open/jag/aidiligenclient/diligen/processor/FieldProcessorTest.java
@@ -1,7 +1,7 @@
 package ca.bc.gov.open.jag.aidiligenclient.diligen.processor;
 
 import ca.bc.gov.open.jag.aidiligenclient.diligen.model.DocumentConfig;
-import ca.bc.gov.open.jag.efilingdiligenclient.api.model.Field;
+import ca.bc.gov.open.jag.aidiligenclient.api.model.Field;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
